Remove default systemd configuration
https://www.freedesktop.org/software/systemd/man/systemd.socket.html

Service=
Specifies the service unit name to activate on incoming traffic. This setting is only allowed for sockets with Accept=no.
It defaults to the service that bears the same name as the socket (with the suffix replaced).
In most cases, it should not be necessary to use this option.
Note that setting this parameter might result in additional dependencies to be added to the unit (see above)

Change-Id: Ia13b59b69ca9cd014d0c4e7ab9e744eb6c0b81ab